--- src/components/Button.jsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-
-class ReactButton extends React.Component{
-    constructor(props){
-        super(props)
-    }
-
-
-}
-export default ReactButton
