+++ src/components/utils/Button.jsx
@@ -1,0 +1,8 @@
+import React from "react";
+
+class ReactButton extends React.Component {
+  constructor(props) {
+    super(props);
+  }
+}
+export default ReactButton;
