+++ src/views/login/logController.jsx
@@ -1,18 +1,17 @@
-import React from 'react';
-import * as Ivar from './Ivar';
-import SignIn from '../../components/SignIn';
-import SignUp from '../../components/SignUp';
-import { Auth } from 'aws-amplify';
+import React from "react";
+import SignIn from "../../components/LogIn/SignIn";
+import SignUp from "../../components/LogIn/SignUp";
+import { Auth } from "aws-amplify";
 
 class logController extends React.Component {
   constructor(props) {
-    super(props)
+    super(props);
 
     this.state = {
       isUser: true,
-      email: '',
-      password: '',
-    }
+      email: "",
+      password: ""
+    };
   }
 
   async signUp(user, password, email) {
@@ -29,19 +28,18 @@ class logController extends React.Component {
   }
 
   async signIn(user, password) {
-    console.log(user, password)
+    console.log(user, password);
     try {
       await Auth.signIn(user, password);
       console.log(Auth.currentAuthenticatedUser());
     } catch (e) {
       alert(e.message);
     }
-  };
+  }
 
   async signOut() {
-    Auth.currentSession()
-      .catch(err => console.log(err));
-  };
+    Auth.currentSession().catch(err => console.log(err));
+  }
 
   async changePassword(oldPassword, newPassword) {
     Auth.currentAuthenticatedUser()
@@ -52,33 +50,22 @@ class logController extends React.Component {
   }
 
   async forgotPassword(user) {
-    Auth.forgotPassword(user)
-      .catch(err => console.log(err));
-  };
+    Auth.forgotPassword(user).catch(err => console.log(err));
+  }
 
   async verifyCurrentUserAttribute() {
-    Auth.verifyCurrentUserAttribute('email')
+    Auth.verifyCurrentUserAttribute("email")
       .then(() => {
         console.log("a verification mail is sent");
       })
       .catch(e => {
         console.log("failed with error", e);
       });
-  };
+  }
 
   render() {
-    return (
-      <>
-        { 
-          this.state.isUser ? (
-            <SignIn />
-          ) : (
-              <SignUp />
-            )
-        }
-      </>
-    )
-  };
-};
+    return <>{this.state.isUser ? <SignIn /> : <SignUp />}</>;
+  }
+}
 
 export default logController;
